Split Statistics.kt into three files, add some comments
