+++ app/src/main/java/com/samco/trackandgraph/statistics/Statistics.kt
@@ -21,21 +21,14 @@ import com.androidplot.Region
 import com.androidplot.util.SeriesUtils
 import com.androidplot.xy.FastXYSeries
 import com.androidplot.xy.RectRegion
-import com.androidplot.xy.StepMode
 import com.samco.trackandgraph.database.TrackAndGraphDatabaseDao
 import com.samco.trackandgraph.database.entity.*
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import kotlinx.coroutines.yield
 import org.threeten.bp.DayOfWeek
 import org.threeten.bp.Duration
 import org.threeten.bp.OffsetDateTime
-import org.threeten.bp.Period
-import org.threeten.bp.temporal.TemporalAdjusters
 import org.threeten.bp.temporal.TemporalAmount
-import org.threeten.bp.temporal.WeekFields
-import java.lang.Exception
-import java.util.*
 import kotlin.math.*
 
 class DataSample(val dataPoints: List<DataPointInterface>)
@@ -96,340 +89,6 @@ private fun getLastTrackedTimeOrNow(
     return listOfNotNull(now, latest).max()!!
 }
 
-/**
- * Add up all data points per plotTotalTime. For example if the plot total time is 1 day and the
- * sample data contains 3 data points {1, 3, 7} all tracked on the same day then the function will
- * return a data sample containing 1 point with the value 11.
- *
- * The currently supported plotTotalTime values are: Duration.ofHours(1), Period.ofDays(1),
- * Period.ofWeeks(1), Period.ofMonths(1), Period.ofYears(1)
- *
- * If sampleDuration is provided then totals will be generated at least as far back as now minus the
- * sampleDuration. However if there is more data before this period then it that data will also be
- * totalled. For clipping see clipDataSample.
- *
- * If end time is provided then totals will be generated at least up to this time. However if there
- * is more data after the end time in the input sample then that data will also be totalled. For
- * clipping see clipDataSample.
- *
- * All notes and labels will be lost in the output data.
- *
- * sampleData.dataPoints should be sorted from oldest timestamp to newest timestamp
- */
-internal suspend fun calculateDurationAccumulatedValues(
-    sampleData: DataSample,
-    featureId: Long,
-    sampleDuration: Duration?,
-    endTime: OffsetDateTime?,
-    plotTotalTime: TemporalAmount,
-    aggPreferences: AggregationWindowPreferences? = null
-): DataSample {
-    return calculateDurationAggregatedValues(sampleData,
-        featureId,
-        sampleDuration,
-        endTime,
-        plotTotalTime,
-        aggPreferences
-    ).sum()
-}
-
-// Contains list of AggregatedDatapoints which did not have a function called on them yet!
-class RawAggregatedDatapoints(private val points: List<AggregatedDataPoint>) {
-    fun sum() = DataSample(
-        this.points.map { it.copy(value = it.parents.sumOf { par -> par.value }) }
-    )
-
-
-    // When using duration based aggregated values, some points can have no parents.
-    // The most intuitive solution is probably to just remove these aggregated points
-    // There might be cases where it can make sense to define a fallback value, e.g. 0
-    fun max() = DataSample(
-        this.points
-            .filter { it.parents.isNotEmpty() }
-            .map { it.copy(value = it.parents.maxOf { par -> par.value }) }
-    )
-
-    fun average() = DataSample(
-        this.points
-            .filter { it.parents.isNotEmpty() }
-            .map { it.copy(value = it.parents.map { par -> par.value }.average()) }
-    )
-}
-
-internal fun DataPointInterface.cutoffTimestampForAggregation(startTimeOfDayOrNull: Duration?): OffsetDateTime {
-    val startTimeOfDay = startTimeOfDayOrNull ?: Duration.ZERO
-    return timestamp - startTimeOfDay
-}
-
-private suspend fun calculateDurationAggregatedValues(
-    sampleData: DataSample,
-    featureId: Long,
-    sampleDuration: Duration?,
-    endTime: OffsetDateTime?,
-    plotTotalTime: TemporalAmount,
-    aggPreferencesOrNull: AggregationWindowPreferences? = null,
-): RawAggregatedDatapoints {
-    val aggPref = aggPreferencesOrNull ?: AggregationWindowPreferences()
-
-    val newData = mutableListOf<AggregatedDataPoint>()
-
-    val firstDataPointTime =
-        sampleData.dataPoints.firstOrNull()?.cutoffTimestampForAggregation(aggPref.startTimeOfDay)
-    val lastDataPointTime =
-        sampleData.dataPoints.lastOrNull()?.cutoffTimestampForAggregation(aggPref.startTimeOfDay)
-
-    val latest = getEndTimeNowOrLatest(lastDataPointTime, endTime)
-    val earliest = getStartTimeOrFirst(firstDataPointTime, latest, endTime, sampleDuration)
-    var currentTimeStamp =
-        findBeginningOfTemporal(earliest, plotTotalTime, aggPref.firstDayOfWeek).minusNanos(1)
-    var index = 0
-    while (currentTimeStamp.isBefore(latest)) {
-        currentTimeStamp = currentTimeStamp.with { ld -> ld.plus(plotTotalTime) }
-        val points = sampleData.dataPoints.drop(index)
-            .takeWhile { dp ->
-                dp.cutoffTimestampForAggregation(aggPref.startTimeOfDay)
-                    .isBefore(currentTimeStamp)
-            }
-        index += points.size
-        newData.add(
-            AggregatedDataPoint(
-            timestamp = currentTimeStamp,
-            value = Double.NaN, // this value gets overwritten when calling a function on RawAggregatedDatapoints
-            featureId = featureId,
-            parents = points))
-        yield()
-    }
-    return RawAggregatedDatapoints(newData)
-}
-
-
-private fun getStartTimeOrFirst(
-    firstDataPointTime: OffsetDateTime?,
-    latest: OffsetDateTime,
-    endTime: OffsetDateTime?,
-    sampleDuration: Duration?
-): OffsetDateTime {
-    val beginningOfDuration = sampleDuration?.let { endTime?.minus(it) }
-    val durationBeforeLatest = sampleDuration?.let { latest.minus(it) }
-    return listOf(
-        firstDataPointTime,
-        beginningOfDuration,
-        durationBeforeLatest,
-        latest
-    ).minBy { t -> t ?: OffsetDateTime.MAX }!!
-}
-
-private fun getEndTimeNowOrLatest(
-    lastDataPointTime: OffsetDateTime?,
-    endTime: OffsetDateTime?
-): OffsetDateTime {
-    val now = OffsetDateTime.now()
-    return when {
-        //last == null && endTime == null -> now
-        endTime == null -> listOf(lastDataPointTime, now).maxBy { t -> t ?: OffsetDateTime.MIN }!!
-        else -> listOf(lastDataPointTime, endTime).maxBy { t -> t ?: OffsetDateTime.MIN }!!
-    }
-}
-
-/**
- * Finds the first ending of temporalAmount before dateTime. For example if temporalAmount is a
- * week then it will find the very end of the sunday before dateTime.
- *
- * temporalAmount supports the use of duration instead of Period.
- * In this case the function will try to approximate the behaviour as closely as possible as though
- * temporalAmount was a period. For example if you pass in a duration of 7 days it will try to find
- * the last day of the week before the week containing dateTime. It is always the largest recognised
- * duration that is used when deciding what the start of the period should be. The recognised durations
- * are:
- *
- *  Duration.ofHours(1)
- *  Duration.ofHours(24)
- *  Duration.ofDays(7)
- *  Duration.ofDays(30)
- *  Duration.ofDays(365 / 4) or 3 months
- *  Duration.ofDays(365 / 2) or 6 months
- *  Duration.ofDays(365) or a year
- *
- */
-
-internal fun getFirstDayOfWeekFromLocale(): DayOfWeek =
-    WeekFields.of(Locale.getDefault()).firstDayOfWeek
-
-internal data class AggregationWindowPreferences(
-    val firstDayOfWeek: DayOfWeek,
-    val startTimeOfDay: Duration
-) {
-    constructor() : this(getFirstDayOfWeekFromLocale(), Duration.ofSeconds(0))
-    constructor(firstDayOfWeek: DayOfWeek) : this(firstDayOfWeek, Duration.ofSeconds(0))
-    constructor(startTimeOfDay: Duration) : this(getFirstDayOfWeekFromLocale(), startTimeOfDay)
-}
-
-internal fun findBeginningOfTemporal(
-    dateTime: OffsetDateTime,
-    temporalAmount: TemporalAmount,
-    firstDayOfWeekOrNull: DayOfWeek? = null
-): OffsetDateTime {
-    // if no firstDayOfWeek is given, use the default one from default constructor
-    val firstDayOfWeek = firstDayOfWeekOrNull ?: AggregationWindowPreferences().firstDayOfWeek
-    return when (temporalAmount) {
-        is Duration -> findBeginningOfDuration(dateTime, temporalAmount, firstDayOfWeek)
-        is Period -> findBeginningOfPeriod(dateTime, temporalAmount, firstDayOfWeek)
-        else -> dateTime
-    }
-}
-
-private fun findBeginningOfDuration(
-    dateTime: OffsetDateTime,
-    duration: Duration,
-    firstDayOfWeek: DayOfWeek
-): OffsetDateTime {
-    val dtHour = dateTime
-        .withMinute(0)
-        .withSecond(0)
-        .withNano(0)
-
-    if (duration <= Duration.ofMinutes(60)) return dtHour
-
-    val dt = dtHour.withHour(0)
-
-    return when {
-        duration <= Duration.ofDays(1) -> dt
-        duration <= Duration.ofDays(7) -> {
-            dt.with(
-                TemporalAdjusters.previousOrSame(
-                    firstDayOfWeek
-                )
-            )
-        }
-        duration <= Duration.ofDays(31) -> dt.withDayOfMonth(1)
-        duration <= Duration.ofDays(ceil(365.0 / 4).toLong()) -> {
-            val month = getQuaterForMonthValue(dt.monthValue)
-            dt.withMonth(month).withDayOfMonth(1)
-        }
-        duration <= Duration.ofDays(ceil(365.0 / 2).toLong()) -> {
-            val month = getBiYearForMonthValue(dt.monthValue)
-            dt.withMonth(month).withDayOfMonth(1).withHour(0)
-        }
-        else -> dt.withDayOfYear(1)
-    }
-}
-
-/**
- * Given a number representing a month in the range 1 to 12, this function will return you the
- * integer value of the month starting the quater of the year containing that month.
- */
-internal fun getQuaterForMonthValue(monthValue: Int) = (3 * ((monthValue - 1) / 3)) + 1
-
-/**
- * Given a number representing a month in the range 1 to 12, this function will return you the
- * integer value of the month starting the bi year containing that month.
- */
-internal fun getBiYearForMonthValue(monthValue: Int) = if (monthValue < 7) 1 else 7
-
-private fun findBeginningOfPeriod(
-    dateTime: OffsetDateTime,
-    period: Period,
-    firstDayOfWeek: DayOfWeek
-): OffsetDateTime {
-    val dt = dateTime.withHour(0)
-        .withMinute(0)
-        .withSecond(0)
-        .withNano(0)
-
-    return when {
-        isPeriodNegativeOrZero(period.minus(Period.ofDays(1))) -> dt
-        isPeriodNegativeOrZero(period.minus(Period.ofWeeks(1))) -> {
-            dt.with(TemporalAdjusters.previousOrSame(firstDayOfWeek))
-        }
-        isPeriodNegativeOrZero(period.minus(Period.ofMonths(1))) -> {
-            dt.withDayOfMonth(1)
-        }
-        isPeriodNegativeOrZero(period.minus(Period.ofMonths(3))) -> {
-            val month = getQuaterForMonthValue(dateTime.monthValue)
-            dt.withMonth(month).withDayOfMonth(1)
-        }
-        isPeriodNegativeOrZero(period.minus(Period.ofMonths(6))) -> {
-            val month = getBiYearForMonthValue(dateTime.monthValue)
-            dt.withMonth(month).withDayOfMonth(1)
-        }
-        else -> dt.withDayOfYear(1)
-    }
-}
-
-private fun isPeriodNegativeOrZero(period: Period) = period.years < 0
-        || (period.years == 0 && period.months < 0)
-        || (period.years == 0 && period.months == 0 && period.days <= 0)
-
-/**
- * Calculate the moving averages of all of the data points given over the moving average duration given.
- * A new DataSample will be returned with one data point for every data point in the input set whose
- * timestamp shall be the same but value will be equal to the average of it and all previous data points
- * within the movingAvDuration.
- *
- * The data points in the input sample are expected to be in date order with the oldest data points
- * earliest in the list
- */
-internal suspend fun calculateMovingAverages(
-    dataSample: DataSample,
-    movingAvgDuration: Duration
-): DataSample {
-    return movingAggregation(dataSample, movingAvgDuration).average()
-}
-
-internal suspend fun movingAggregation(
-    dataSample: DataSample,
-    movingAggDuration: Duration
-): RawAggregatedDatapoints {
-    /*
-    Maybe sometime in the future it is desired to know here whether the input is aggregated using a
-    daily+ totals aggregator, with the intent to treat data that is *not* aggregated more lenient
-    when it comes to the averaging window. E.g. when i use the daily averaging window and track data
-    at 13:01 one day, 13:00 the next day and 13:02 the day after that, the second entry gets averaged,
-    but the third does not etc.
-    One way to check whether the input is aggregated is to check if the time-differences
-    between all consecutive datasamples are exactly the same which likely is only the case if a computer
-    aggregated it.
-    The code below implements this approach, but is commented out since a use for this would have to
-    be discussed first.
-
-    val input_is_aggregated_total_data = when (dataPoints.size) {
-            1, 2 -> false // return false for 2, bc there is no comparison if there is only one time-difference
-            else -> {
-                val inital_time_difference = Duration.between(
-                        dataPoints[0].timestamp,
-                        dataPoints[1].timestamp
-                )
-                (1..dataPoints.size-1).map( { Duration.between(
-                        dataPoints[it-1].timestamp,
-                        dataPoints[it].timestamp) }).filter( { it != inital_time_difference }).size == 0
-            }
-    }
-    */
-    val movingAggregationPointsRaw = mutableListOf<AggregatedDataPoint>()
-    val dataPointsReversed = dataSample.dataPoints.reversed()
-
-    for ( (index, current) in dataPointsReversed.mapIndexed { idx, point -> Pair(idx, point) }) {
-        yield()
-        val parents = dataPointsReversed.drop(index)
-            .takeWhile { dp ->
-                Duration.between(dp.timestamp, current.timestamp) < movingAggDuration
-            }
-
-        movingAggregationPointsRaw.add(
-            0,
-            AggregatedDataPoint(
-                current.timestamp,
-                current.featureId,
-                value = Double.NaN,
-                label = current.label,
-                note = current.note,
-                parents = parents
-            )
-        )
-    }
-
-    return RawAggregatedDatapoints(movingAggregationPointsRaw)
-}
 
 /**
  * Return all the data points in the sample that lie within the sampleDuration leading up to the endTime.
@@ -687,185 +346,4 @@ internal fun getLargestBin(bins: List<List<Double>>?): Double? {
         ?.max()
 }
 
-
-/**
- * Code to calculate nice y-intervals
- */
-// There are two possible interval classes. The reason is that we want to make a large amount of
-// objects of the first class, which then will get filtered. Then we only have to do the more complex
-// calculations for the actually reasonable intervals
-data class PossibleIntervalProto(
-    val interval: Double, val preferred: Boolean,
-    val base: Double
-)
-
-data class PossibleInterval(
-    val interval: Double, val preferred: Boolean,
-    val n_lines: Int, val percentage_range_used: Double,
-    val bounds_min: Double, val bounds_max: Double,
-    val base: Double
-)
-
-data class YAxisParameters(
-    val step_mode: StepMode, val n_intervals: Double,
-    val bounds_min: Double, val bounds_max: Double,
-)
-
-private fun finishProto(
-    proto: PossibleIntervalProto,
-    y_min: Double,
-    y_max: Double,
-    fixedBounds: Boolean
-): PossibleInterval? {
-    return when (fixedBounds) {
-        true -> finishProtoFixedBounds(proto, y_min, y_max)
-        false -> finishProtoDynamicBounds(proto, y_min, y_max)
-    }
-}
-
-private fun finishProtoDynamicBounds(
-    proto: PossibleIntervalProto,
-    y_min: Double,
-    y_max: Double
-): PossibleInterval {
-    val y_range = y_max - y_min
-    var bounds_min = floor(y_min / proto.interval) * proto.interval
-    var bounds_max = ceil(y_max / proto.interval) * proto.interval
-
-    // check whether there is a large gap between bounds and data and if its larger
-    // than half the base on both sides, remove it
-    // it can only be this large if the divisor is 1
-    val offset = proto.base / 2
-    if (y_min - bounds_min >= offset && bounds_max - y_max >= offset) {
-        bounds_min += offset
-        bounds_max -= offset
-    }
-
-    val bounds_range = bounds_max - bounds_min
-    val n_lines = round(bounds_range / proto.interval).toInt() + 1
-    val percentage_range_used = y_range / bounds_range
-
-    return PossibleInterval(
-        proto.interval, proto.preferred,
-        n_lines, percentage_range_used,
-        bounds_min, bounds_max,
-        proto.base
-    )
-}
-
-private fun finishProtoFixedBounds(
-    proto: PossibleIntervalProto,
-    y_min: Double,
-    y_max: Double
-): PossibleInterval? {
-    val y_range = y_max - y_min
-    // if we can't evenly divide the y_range we can't do anything, so return a bad dummy interval, which will get filtered out
-    // this is a little more complex than it should be because of floating point errors
-    if (y_range.div(proto.interval)
-            .rem(1) != 0.0
-    ) return null //PossibleInterval(0.0, false, 999, 0.0, 0.0, 0.0, 1.0)
-
-    return PossibleInterval(
-        proto.interval, proto.preferred, round(y_range / proto.interval).toInt() + 1,
-        1.0, y_min, y_max, proto.base
-    )
-}
-
-
-fun getYParameters(
-    y_min: Double, y_max: Double, time_data: Boolean,
-    fixedBounds: Boolean,
-    throw_exc_if_non_found: Boolean = false
-): YAxisParameters {
-
-
-    val parameters =
-        getYParametersInternal(y_min, y_max, time_data, fixedBounds, throw_exc_if_non_found)
-    if (parameters != null) {
-        return YAxisParameters(
-            StepMode.SUBDIVIDE,
-            parameters.n_lines.toDouble(),
-            parameters.bounds_min,
-            parameters.bounds_max
-        )
-    }
-
-    // fallback if we don't find any solution. Gets used when all our solutions use to little of the range.
-    // this was the default behavior before this new algorithm existed
-    val fallback = YAxisParameters(StepMode.SUBDIVIDE, 11.0, y_min, y_max)
-    return fallback
-
-}
-
-fun getYParametersInternal(
-    y_min: Double, y_max: Double, time_data: Boolean,
-    fixedBounds: Boolean, throw_exc_if_non_found: Boolean
-): PossibleInterval? {
-    val MIN_INTERVALS = 6
-    val MAX_INTERVALS = 12
-    val MIN_USED_RANGE_STEPS = listOf<Double>(0.849, 0.79)//, 0.749, 0.7)
-
-    val y_range = y_max - y_min
-
-    val (base, preferred_divisors, all_divisors) = when (time_data) {
-        false -> Triple(
-            10, setOf(1, 2, 4, 5),
-            setOf(1, 2, 3, 4, 5, 8)
-        )
-        true -> Triple(
-            60, setOf(1, 2, 3, 4, 6, 12, 24),
-            setOf(1, 2, 3, 4, 6, 12, 24, 30)
-        )
-    }
-
-    // We want to treat a y_range of 60 the same way as a range of 600 or 6, just times 10 or 0.1
-    val norm_exponent = round(log(y_range, base.toDouble()))
-    val normed_base = Math.pow(base.toDouble(), norm_exponent.toDouble())
-
-    val reasonable_intervals = all_divisors
-        .flatMap { div ->
-            (-1..1).map { exp_offset ->
-                PossibleIntervalProto(
-                    normed_base * Math.pow(base.toDouble(), exp_offset.toDouble()) / div,
-                    div in preferred_divisors,
-                    base = normed_base * Math.pow(base.toDouble(), exp_offset.toDouble())
-                )
-            }
-        }
-        // so after we generate a lot of possible interval prototypes, in this first filter we get rid of all the
-        // ones who just are not plausible at all. We'll do a second check with the exact data further below
-        .filter { interval ->
-            ceil(y_range / interval.interval) >= MIN_INTERVALS - 1 &&
-                    ceil(y_range / interval.interval) <= MAX_INTERVALS + 1
-        }
-        .map { proto -> finishProto(proto, y_min, y_max, fixedBounds = fixedBounds) }
-        .filterNotNull()
-        .filter { interval ->
-            interval.n_lines >= MIN_INTERVALS &&
-                    interval.n_lines <= MAX_INTERVALS
-        }
-
-    if (reasonable_intervals.isEmpty()) {
-        if (throw_exc_if_non_found) throw Exception("No solution found! No intervals passed the initial filtering. ymin: $y_min, ymax: $y_max")
-        return null
-
-    }
-
-    for (MIN_USED_RANGE in MIN_USED_RANGE_STEPS) {
-        // gradually have lower expectations in our output
-
-        for (pref in listOf(true, false)) {
-            // prefer the 'preferred' intervals when it comes to each used_range step
-            reasonable_intervals
-                .filter { it.preferred == pref }
-                .filter { it.percentage_range_used >= MIN_USED_RANGE }
-                .sortedByDescending { it.interval }   // prefer larger intervals
-                .forEach { return it }  // returns the first element, if there is one
-        }
-    }
-
-    if (throw_exc_if_non_found) throw Exception("No solution found! No intervals passed the range_used check. ymin: $y_min, ymax: $y_max")
-    return null
-
-}
